+++ include/zenoh-pico/link/config/serial.h
@@ -41,7 +41,7 @@
 // #define SERIAL_CONFIG_TOUT_STR         "tout"
 
 #define SERIAL_CONFIG_MAPPING_BUILD            \
-    int argc = 1;                              \
+    uint8_t argc = 1;                              \
     _z_str_intmapping_t args[argc];            \
     args[0]._key = SERIAL_CONFIG_BAUDRATE_KEY; \
     args[0]._str = SERIAL_CONFIG_BAUDRATE_STR;
